+++ UIComponents.Abstractions/Interfaces/ValidationRules/IDefaultCheckValidationErrors.cs
@@ -4,5 +4,5 @@ namespace UIComponents.Abstractions.Interfaces.ValidationRules;
 
 public interface IUICDefaultCheckValidationErrors<TValidator> where TValidator: IUICPropertyValidationRule
 {
-    Task<ValidationRuleResult> DefaultValidationErrors(TValidator validator, PropertyInfo propertyInfo, object obj);
+    Task<ValidationRuleResult> DefaultValidationErrors(TValidator validator, PropertyInfo propertyInfo, object obj, CancellationToken cancellationToken = default);
 }