+++ UIComponents.Generators/Validators/CustomValidators/CustomValidatorPropertyRequired.cs
@@ -6,16 +6,16 @@ namespace UIComponents.Generators.Validators.CustomValidators;
 public class CustomValidatorPropertyRequired : IUICPropertyValidationRuleRequired
 {
 
-    public Func<PropertyInfo,object,Task<bool>> IsRequiredFunc { get; set; }
+    public Func<PropertyInfo,object, CancellationToken,Task<bool>> IsRequiredFunc { get; set; }
 
     public Type? PropertyType => typeof(object);
 
 
-    public Task<bool> IsRequired(PropertyInfo propertyInfo, object obj)
+    public Task<bool> IsRequired(PropertyInfo propertyInfo, object obj, CancellationToken cancellationToken = default)
     {
         if (IsRequiredFunc == null)
             throw new ArgumentNullException(nameof(IsRequiredFunc));
 
-        return IsRequiredFunc(propertyInfo, obj);
+        return IsRequiredFunc(propertyInfo, obj, cancellationToken);
     }
 }